+++ src/components/date/date.stories.tsx
@@ -1,17 +1,21 @@
 import React, { ChangeEvent, useState } from 'react';
 import { DateInput } from './index';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
-import { dateInitiValue, DateValueType } from '../../hooks/useCalender';
+import { DateValueType } from '../../hooks/useCalender';
 
 export default {
   title: 'component/date',
   component: DateInput,
 } as ComponentMeta<typeof DateInput>;
 
+const getDate = new Date();
+
 export const Template: ComponentStory<typeof DateInput> = (arg) => {
-  const [state, setState] = useState<DateValueType>(
-    dateInitiValue('includeDay')
-  );
+  const [state, setState] = useState<DateValueType>({
+    year: getDate.getFullYear(),
+    month: getDate.getMonth(),
+    day: getDate.getDate(),
+  });
 
   const onDateClick = (value: DateValueType) => {
     setState(value);
